Refactor env script handling in Tailscale settings
Moves environment script content to top-level constants and refines logic for creating, updating, or removing /etc/profile.d/tailscale-env.sh. Ensures the script is only managed when necessary and restarts Tailscale service after changes.

+++ luci-app-tailscale-community/root/usr/share/rpcd/ucode/tailscale.uc
@@ -6,6 +6,15 @@ import { access, popen, readfile, writefile, unlink } from 'fs';
 import { cursor } from 'uci';
 
 const uci = cursor();
+const env_script_path = "/etc/profile.d/tailscale-env.sh";
+const ori_env_script_content = `#!/bin/sh
+# This script is managed by luci-app-tailscale-community.
+uci_get_state() { uci get tailscale.settings."$1" 2>/dev/null; }
+if [ "$(uci_get_state daemon_reduce_memory)" = "1" ]; then export GOGC=10; fi
+TS_MTU=$(uci_get_state daemon_mtu)
+if [ -n "$TS_MTU" ]; then export TS_DEBUG_MTU="$TS_MTU"; fi
+`;
+const env_script_content = replace(ori_env_script_content, /\r/g, '');
 
 function exec(command) {
     let stdout_content = '';
@@ -199,24 +208,17 @@ methods.set_settings = {
 
         let new_mtu = form_data.daemon_mtu || "";
         let new_reduce_mem = form_data.daemon_reduce_memory || "0";
-        if (new_mtu != null || new_mtu != '0' || new_reduce_mem != 0) {
-            try{mkdir('/etc/profile.d');} catch (e) { }
-            const env_script_path = "/etc/profile.d/tailscale-env.sh";
-            const env_script_content = `#!/bin/sh
-# This script is managed by luci-app-tailscale-community.
-uci_get_state() { uci get tailscale.settings."$1" 2>/dev/null; }
-if [ "$(uci_get_state daemon_reduce_memory)" = "1" ]; then export GOGC=10; fi
-TS_MTU=$(uci_get_state daemon_mtu)
-if [ -n "$TS_MTU" ]; then export TS_DEBUG_MTU="$TS_MTU"; fi
-`;
-            const clean_env_script_content = replace(env_script_content, /\r/g, '');
-            if (new_mtu !== "" || new_reduce_mem === "1") {
-                writefile(env_script_path, clean_env_script_content);
-                exec('chmod 755 '+env_script_path);
-            } else {
-                unlink(env_script_path);
+        if (access('/etc/profile.d/tailscale-env.sh')==false) {
+            if (new_mtu != null || new_mtu != '0' || new_reduce_mem != 0) {
+                try{mkdir('/etc/profile.d');} catch (e) { }
+                if (new_mtu !== "" || new_reduce_mem === "1") {
+                    writefile(env_script_path, env_script_content);
+                    exec('chmod 755 '+env_script_path);
+                } else {
+                    unlink(env_script_path);
+                }
+                popen('/bin/sh -c /etc/init.d/tailscale restart &');
             }
-            popen('/bin/sh -c /etc/init.d/tailscale restart &');
         }
         return { success: true };
     }
